ai4realnet_orchestrators/power_grid/framework/requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiohttp-cors==0.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blosc2==2.7.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
ConfigSpace==1.2.1
contourpy==1.3.1
cycler==0.12.1
deap==1.4.2
debugpy==1.8.13
decorator==5.2.1
deepdiff==8.4.2
defopt==6.4.0
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
dm-tree==0.1.9
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
Grid2Op==1.10.5
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
gymnasium-notices==0.0.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython  
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
joypy==0.2.6
json-tricks==3.17.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.12.0
kiwisolver==1.4.8
l2rpn_baselines==0.8.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.5.1
lightning-utilities==0.14.2
LightSim2Grid==0.10.2
llvmlite==0.44.0
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.3
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.0
multiprocess==0.70.17
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numexpr==2.10.2
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opt_einsum==3.4.0
orderly-set==5.3.0
ormsgpack==1.7.0
overrides==7.7.0
packaging==24.2
pandapower==2.14.11
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathos==0.3.3
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pockets==0.9.1
pox==0.3.5
ppft==1.7.6.9
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
PythonWebHDFS==0.2.3
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.25.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
sb3_contrib==2.6.0
schema==0.7.7
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
Shimmy==2.0.0
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
sphinxcontrib-napoleon==0.7
stack-data==0.6.3
statsmodels==0.14.4
sympy>=1.13.1
tables==3.10.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-estimator==2.12.0
tensorflow-probability==0.20.1
termcolor==3.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tomli==2.2.1
torchmetrics==1.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==2.13.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ubelt==1.3.7
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.31.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.14.1
yarg==0.1.9
yarl==1.18.3
flatland-rl == 4.2.2